pglock: Increase concurrency by moving RVM outside main Tx
Due to the use of the Serializable isolation level for the transaction when many
locks are being established concurrently from different hosts, even if the locks
are on different keys, they can conflict due to the shared use of `nextval()` on
the table sequence. This change moves the call to `nextval()` outside the main
trainsaction. The call is still atomic and each client will receive a unique
token, but now only clients trying to set the same lock will have potential
conflicts.

Co-authored-by: name <u@cirello.eu>